Move worker queue information to a separate key
Resque workers are currently identified by hostname:pid:queues. However,
many of our workers are using very large queue lists, resulting in
worker ids that are 500 bytes or more. Instead of storing the list of
queues in the worker id, move the queues to a separate key.

The hope is that this will  reduce contention on the `resque:workers`
set by making the keys smaller, in hopes that smaller keys will result
in faster performance both for adding/removing members as well as the
overall size of the data required to send the entire list of workers via
SMEMBERS.

The list of queues is preserved in a separate key for use with
debugging tools like `gh-resque`.